Create Data dig project Marvels heroes and Villains
This is my project data dig on Marvel super heroes and villains. I queried the data to find average heights from all the characters. I then sorted the table to show me some conditions I was interested about and sorted the table by most and least popular characters. I then queried the table for the 10 most popular characters and further dug for data on the female characters that are most popular.