Create Project Famous people
In this project, I created a database of all 20 Formula 1 drivers and some data on them. I used an inner join to aggregate data on the drivers with the most wins and the country they race for. I was also interested in data regarding birth years and sorted the data accordingly.